requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
langchain==0.1.20
langchain-community==0.3.27
langchain-openai==0.1.6
openai==1.97.1
python-dotenv==1.1.1
